Use global math functions instead of std

+++ Plugin/Source/Processors/Loss Effects/LossEffects.cpp
@@ -16,9 +16,9 @@ void LossEffects::init (float sampleRate, float speed, float spacing, float thic
         const auto freq = ((float) k * binWidth) + (binWidth / 2.0f);
         const auto waveNumber = MathConstants<float>::twoPi * freq / speedMetric;
 
-        const auto spacingLoss = std::exp (-1.0f * waveNumber * spacing);
-        const auto gapLoss = std::sin (waveNumber * gap / 2.0f) / (waveNumber * gap / 2.0f);
-        const auto thicknessLoss = (1.0f - std::exp (-waveNumber * thickness)) / (waveNumber * thickness);
+        const auto spacingLoss = expf (-1.0f * waveNumber * spacing);
+        const auto gapLoss = sinf (waveNumber * gap / 2.0f) / (waveNumber * gap / 2.0f);
+        const auto thicknessLoss = (1.0f - expf (-waveNumber * thickness)) / (waveNumber * thickness);
 
         H[k] = spacingLoss * gapLoss * thicknessLoss;
         H[order - k - 1] = H[k];
@@ -29,7 +29,7 @@ void LossEffects::init (float sampleRate, float speed, float spacing, float thic
     {
         h[n] = 0;
         for (int k = 0; k < order; k++)
-            h[n] += H[k] * std::cos (MathConstants<float>::twoPi * (float) k * (float) n / (float) order);
+            h[n] += H[k] * cosf (MathConstants<float>::twoPi * (float) k * (float) n / (float) order);
 
         h[n] /= (float) order;
     }
